refactor: Replace JenjangSekolah and Kelas with BentukPendidikan, update related models, controllers, and migrations

+++ database/migrations/2025_04_14_134319_create_mata_pelajarans_table.php
@@ -10,7 +10,7 @@ return new class extends Migration {
         Schema::create('mata_pelajarans', function (Blueprint $table) {
             $table->id();
             $table->string('nama_mapel');
-            $table->float('bobot');
+            $table->integer('bobot');
             $table->timestamps();
             $table->softDeletes();
         });
